# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type TransactionCoin = typeof TransactionCoin[keyof typeof TransactionCoin];


export const TransactionCoin = {
  BTC: 'BTC',
  ETH: 'ETH',
} as const;

export type TransactionSource = typeof TransactionSource[keyof typeof TransactionSource];


export const TransactionSource = {
  manual: 'manual',
  indodax_sync: 'indodax_sync',
} as const;

export interface Transaction {
  id: string;
  date: string;
  type: TransactionType;
  coin: TransactionCoin;
  jumlahBeli: number;
  fee: number;
  hargaBeliPerCoin: number;
  source: TransactionSource;
  /** @nullable */
  pair?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  tradeId?: string | null;
  createdAt: string;
}

export type TransactionEntryType = typeof TransactionEntryType[keyof typeof TransactionEntryType];


export const TransactionEntryType = {
  BUY: 'BUY',
  SELL: 'SELL',
} as const;

export type TransactionEntryCoin = typeof TransactionEntryCoin[keyof typeof TransactionEntryCoin];


export const TransactionEntryCoin = {
  BTC: 'BTC',
  ETH: 'ETH',
} as const;

export type TransactionEntrySource = typeof TransactionEntrySource[keyof typeof TransactionEntrySource];


export const TransactionEntrySource = {
  manual: 'manual',
  indodax_sync: 'indodax_sync',
} as const;

export interface TransactionEntry {
  id: string;
  date: string;
  type: TransactionEntryType;
  coin: TransactionEntryCoin;
  jumlahBeli: number;
  fee: number;
  hargaBeliPerCoin: number;
  source: TransactionEntrySource;
  pair?: string;
  orderId?: string;
  tradeId?: string;
}

export interface TransactionImportInput {
  transactions: TransactionEntry[];
}

export interface TransactionImportResult {
  imported: number;
  skipped: number;
  transactions: Transaction[];
}