# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction persistence (currently used for Indodax Trade History import)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all stored transactions
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a single transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionEntry"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /transactions/import:
    post:
      operationId: importTransactions
      tags: [transactions]
      summary: Import a batch of transactions (idempotent by id)
      description: >
        Inserts the given transactions, skipping any whose id already exists
        (e.g. re-syncing Indodax Trade History does not create duplicates).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionImportResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Transaction:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        type:
          type: string
          enum: [BUY, SELL]
        coin:
          type: string
          enum: [BTC, ETH]
        jumlahBeli:
          type: number
        fee:
          type: number
        hargaBeliPerCoin:
          type: number
        source:
          type: string
          enum: [manual, indodax_sync]
        pair:
          type: ["string", "null"]
        orderId:
          type: ["string", "null"]
        tradeId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - type
        - coin
        - jumlahBeli
        - fee
        - hargaBeliPerCoin
        - source
        - createdAt
    TransactionEntry:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        type:
          type: string
          enum: [BUY, SELL]
        coin:
          type: string
          enum: [BTC, ETH]
        jumlahBeli:
          type: number
        fee:
          type: number
        hargaBeliPerCoin:
          type: number
        source:
          type: string
          enum: [manual, indodax_sync]
        pair:
          type: string
        orderId:
          type: string
        tradeId:
          type: string
      required:
        - id
        - date
        - type
        - coin
        - jumlahBeli
        - fee
        - hargaBeliPerCoin
        - source
    TransactionImportInput:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionEntry"
      required:
        - transactions
    TransactionImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - imported
        - skipped
        - transactions