# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionCoin } from './transactionCoin';
import type { TransactionSource } from './transactionSource';
import type { TransactionType } from './transactionType';

export interface Transaction {
  id: string;
  date: Date;
  type: TransactionType;
  coin: TransactionCoin;
  jumlahBeli: number;
  fee: number;
  hargaBeliPerCoin: number;
  source: TransactionSource;
  /** @nullable */
  pair?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  tradeId?: string | null;
  createdAt: Date;
}