# lib/api-zod/src/generated/types/transactionEntry.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionEntryCoin } from './transactionEntryCoin';
import type { TransactionEntrySource } from './transactionEntrySource';
import type { TransactionEntryType } from './transactionEntryType';

export interface TransactionEntry {
  id: string;
  date: Date;
  type: TransactionEntryType;
  coin: TransactionEntryCoin;
  jumlahBeli: number;
  fee: number;
  hargaBeliPerCoin: number;
  source: TransactionEntrySource;
  pair?: string;
  orderId?: string;
  tradeId?: string;
}