# lib/api-zod/src/generated/types/transactionImportInput.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionEntry } from './transactionEntry';

export interface TransactionImportInput {
  transactions: TransactionEntry[];
}