# lib/api-zod/src/generated/types/transactionImportResult.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface TransactionImportResult {
  imported: number;
  skipped: number;
  transactions: Transaction[];
}